requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.1
huggingface_hub>=0.24.0
Pillow>=10.0.0
